requirements.txt
fastapi
ultralytics
numpy
opencv-python
pandas
pvlib
pillow
requests
python-dotenv
python-multipart
